requirements.txt
python-zcashlib
mnemonic
Pillow
qrcode
